feat: add fibonacci algo page
